+++ a_test.go
@@ -4,10 +4,107 @@ import (
 	"context"
 	"io"
 	"io/ioutil"
+	"os"
+	"strings"
 	"testing"
 
 	"github.com/goyek/goyek/v2"
 )
+
+func TestA_Cleanup(t *testing.T) {
+	out := &strings.Builder{}
+
+	got := goyek.NewRunner(func(a *goyek.A) {
+		a.Cleanup(func() {
+			a.Cleanup(func() {
+				a.Log("5")
+				panic("second panic")
+			})
+			a.Cleanup(func() {
+				a.Log("4")
+			})
+			a.Log("3")
+			panic("first panic")
+		})
+		a.Log("1")
+		a.Cleanup(func() {
+			a.Log("2")
+		})
+	})(goyek.Input{Logger: &goyek.FmtLogger{}, Output: out})
+
+	assertEqual(t, got.Status, goyek.StatusFailed, "shoud return proper status")
+	assertEqual(t, got.PanicValue, "first panic", "shoud return proper panic value")
+	assertContains(t, out, "1\n2\n3\n4\n5", "should call cleanup funcs in LIFO order")
+}
+
+func TestA_Cleanup_when_action_panics(t *testing.T) {
+	out := &strings.Builder{}
+
+	got := goyek.NewRunner(func(a *goyek.A) {
+		a.Cleanup(func() {
+			a.Cleanup(func() {
+				a.Log("5")
+				panic("second panic")
+			})
+			a.Cleanup(func() {
+				a.Log("4")
+			})
+			a.Log("3")
+			panic("first panic")
+		})
+		a.Log("1")
+		a.Cleanup(func() {
+			a.Log("2")
+		})
+		panic("action panic")
+	})(goyek.Input{Logger: &goyek.FmtLogger{}, Output: out})
+
+	assertEqual(t, got.Status, goyek.StatusFailed, "shoud return proper status")
+	assertEqual(t, got.PanicValue, "action panic", "shoud return proper panic value")
+	assertContains(t, out, "1\n2\n3\n4\n5", "should call cleanup funcs in LIFO order")
+}
+
+func TestA_Cleanup_Fail(t *testing.T) {
+	got := goyek.NewRunner(func(a *goyek.A) {
+		a.Cleanup(func() {
+			a.Fail()
+		})
+	})(goyek.Input{})
+
+	assertEqual(t, got.Status, goyek.StatusFailed, "shoud return proper status")
+}
+
+func TestA_Setenv(t *testing.T) {
+	key := "GOYEK_TEST_ENV"
+	val := "1"
+	goyek.NewRunner(func(a *goyek.A) {
+		a.Setenv(key, val)
+
+		got := os.Getenv(key)
+		assertEqual(t, got, val, "should set the value")
+	})(goyek.Input{})
+
+	got := os.Getenv(key)
+	assertEqual(t, got, "", "should restore the value after the action")
+}
+
+func TestA_Setenv_restore(t *testing.T) {
+	key := "GOYEK_TEST_ENV"
+	prev := "0"
+	val := "1"
+	os.Setenv(key, prev)   //nolint:errcheck,gosec // should never happen
+	defer os.Unsetenv(key) //nolint:errcheck // should never happen
+
+	goyek.NewRunner(func(a *goyek.A) {
+		a.Setenv(key, val)
+
+		got := os.Getenv(key)
+		assertEqual(t, got, val, "should set the value")
+	})(goyek.Input{})
+
+	got := os.Getenv(key)
+	assertEqual(t, got, prev, "should restore the value after the action")
+}
 
 func TestA_uses_Logger_dynamic_interface(t *testing.T) {
 	testCases := []struct {