+++ CHANGELOG.md
@@ -11,6 +11,8 @@ as well as to [Module version numbering](https://go.dev/doc/modules/version-numb
 ### Added
 
 - Add `A.Cleanup` method that registers an action's cleanup function.
+- Add `A.Setenv` method that sets the environment variable
+  and returns the previous value during cleanup.
 
 <!-- markdownlint-disable-next-line line-length -->
 ## [2.0.0-rc.9](https://github.com/goyek/goyek/compare/v2.0.0-rc.8...v2.0.0-rc.9) - 2022-11-06